requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
dask==2021.8.1
Django==3.2.7
django-debug-toolbar==3.2.2
django-environ==0.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
fsspec==2021.8.1
iniconfig==1.1.1
locket==0.2.1
memory-profiler==0.58.0
numpy==1.21.2
nvidia-ml-py==11.450.51
packaging==21.0
pandas==1.3.2
partd==1.2.0
pluggy==1.0.0
postgres==3.0.0
psutil==5.8.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
rich==10.9.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
toolz==0.11.1